+++ demos/src/Commands/InsertContent/Vue/index.vue
@@ -94,7 +94,7 @@ export default {
 
 <style lang="scss">
 /* Basic editor styles */
-.ProseMirror {
+.tiptap {
   > * + * {
     margin-top: 0.75em;
   }