+++ demos/src/Experiments/Commands/Vue/index.vue
@@ -45,7 +45,7 @@ export default {
 </script>
 
 <style lang="scss">
-.ProseMirror {
+.tiptap {
   > * + * {
     margin-top: 0.75em;
   }