+++ demos/src/Experiments/DestroyingEditor/Vue/index.vue
@@ -41,7 +41,7 @@ export default {
 </script>
 
 <style lang="scss">
-.ProseMirror {
+.tiptap {
   > * + * {
     margin-top: 0.75em;
   }