+++ demos/src/Extensions/Color/React/index.spec.js
@@ -4,10 +4,10 @@ context('/src/Extensions/Color/React/', () => {
   })
 
   beforeEach(() => {
-    cy.get('.ProseMirror').then(([{ editor }]) => {
+    cy.get('.tiptap').then(([{ editor }]) => {
       editor.commands.setContent('<p>Example Text</p>')
     })
-    cy.get('.ProseMirror').type('{selectall}')
+    cy.get('.tiptap').type('{selectall}')
   })
 
   it('should set the color of the selected text', () => {
@@ -16,23 +16,23 @@ context('/src/Extensions/Color/React/', () => {
       .click()
       .should('have.class', 'is-active')
 
-    cy.get('.ProseMirror').find('span').should('have.attr', 'style', 'color: #958DF1')
+    cy.get('.tiptap').find('span').should('have.attr', 'style', 'color: #958DF1')
   })
 
   it('should remove the color of the selected text', () => {
     cy.get('button:first').click()
 
-    cy.get('.ProseMirror span').should('exist')
+    cy.get('.tiptap span').should('exist')
 
     cy.get('button:last').click()
 
-    cy.get('.ProseMirror span').should('not.exist')
+    cy.get('.tiptap span').should('not.exist')
   })
 
   it('should change text color with color picker', () => {
     cy.get('input[type=color]').invoke('val', '#ff0000').trigger('input')
 
-    cy.get('.ProseMirror').find('span').should('have.attr', 'style', 'color: #ff0000')
+    cy.get('.tiptap').find('span').should('have.attr', 'style', 'color: #ff0000')
   })
 
   it('should match text and color picker color values', () => {