+++ demos/src/Extensions/Focus/React/index.spec.js
@@ -4,6 +4,6 @@ context('/src/Extensions/Focus/React/', () => {
   })
 
   it('should have class', () => {
-    cy.get('.ProseMirror p:first').should('have.class', 'has-focus')
+    cy.get('.tiptap p:first').should('have.class', 'has-focus')
   })
 })