+++ demos/src/Marks/Subscript/Vue/index.spec.js
@@ -4,21 +4,21 @@ context('/src/Marks/Subscript/Vue/', () => {
   })
 
   beforeEach(() => {
-    cy.get('.ProseMirror').then(([{ editor }]) => {
+    cy.get('.tiptap').then(([{ editor }]) => {
       editor.commands.setContent('<p>Example Text</p>')
-      cy.get('.ProseMirror').type('{selectall}')
+      cy.get('.tiptap').type('{selectall}')
     })
   })
 
   it('should transform inline style to sub tags', () => {
-    cy.get('.ProseMirror').then(([{ editor }]) => {
+    cy.get('.tiptap').then(([{ editor }]) => {
       editor.commands.setContent('<p><span style="vertical-align: sub">Example Text</span></p>')
       expect(editor.getHTML()).to.eq('<p><sub>Example Text</sub></p>')
     })
   })
 
   it('sould omit inline style with a different vertical align', () => {
-    cy.get('.ProseMirror').then(([{ editor }]) => {
+    cy.get('.tiptap').then(([{ editor }]) => {
       editor.commands.setContent('<p><b style="vertical-align: middle">Example Text</b></p>')
       expect(editor.getHTML()).to.eq('<p>Example Text</p>')
     })
@@ -28,15 +28,15 @@ context('/src/Marks/Subscript/Vue/', () => {
     cy.get('button:first')
       .click()
 
-    cy.get('.ProseMirror')
+    cy.get('.tiptap')
       .find('sub')
       .should('contain', 'Example Text')
   })
 
   it('the button should toggle the selected text bold', () => {
     cy.get('button:first').click()
-    cy.get('.ProseMirror').type('{selectall}')
+    cy.get('.tiptap').type('{selectall}')
     cy.get('button:first').click()
-    cy.get('.ProseMirror sub').should('not.exist')
+    cy.get('.tiptap sub').should('not.exist')
   })
 })